Error handling fixed for commands, Long string splitting for some commands that could potentially return large data sets. Store directory moved, split_at_separator method moved to utils
